model/requirements.txt
catboost==1.2.8
category_encoders==2.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
graphviz==0.20.3
idna==3.10
imbalanced-learn==0.13.0
joblib==1.5.0
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
matplotlib==3.10.3
narwhals==1.40.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotly==6.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.7.1
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
statsmodels==0.14.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
xgboost==3.0.1
